refactor(editor-editing-panel): inline bound sections into settings tab
Fold bound section rendering into SettingsTab and SettingsControl,
removing BoundSettingsSection and TransformableObjectBridge.

Ref: ED-24051
