+++ packages/packages/core/editor-editing-panel/src/controls-registry/settings-field.tsx
@@ -53,7 +53,6 @@ export const SettingsField = ( { bind, children, propDisplayName }: SettingsFiel
 
 		const settingsWithDefaults = getElementSettingsWithDefaults( propsSchema, currentElementSettings );
 		const settings = getUpdatedValues( newValue, dependents, propsSchema, settingsWithDefaults, elementId );
-
 		if ( withHistory ) {
 			undoableUpdateElementProp( settings );
 		} else {
